stock-insight/requirements.txt
yfinance==0.2.36
twelvedata==1.0.0
finnhub-python==2.4.10
pandas==2.1.4
numpy==1.26.3
pandas-ta==0.3.14b
plotly==5.18.0
jinja2==3.1.3
python-dotenv==1.0.1
requests==2.31.0